[orchestrator] move orchestrator package
