+++ pkg/clusteragent/orchestrator/transformer.go
@@ -12,8 +12,7 @@ import (
 	"strings"
 
 	model "github.com/DataDog/agent-payload/process"
-	"github.com/DataDog/datadog-agent/pkg/process/util/orchestrator"
-
+	"github.com/DataDog/datadog-agent/pkg/orchestrator"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
